Refactor separate prefix handling into prefix_extension
